pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-gpt-bot"
version = "0.1.0"
description = "Private Telegram bot backed by the OpenAI Responses API."
requires-python = ">=3.11"
dependencies = [
    "openai>=1.79.0",
    "python-telegram-bot>=21.10,<22.0",
    "SQLAlchemy>=2.0.36,<3.0",
    "python-dotenv>=1.0.1,<2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
